pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tootles"
authors = [{ name="Tootles Contributors", email="contributors@tootles.dev" }]
description = "Modern Textual-based Mastodon client with web UI parity"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
dynamic = ["version"]

classifiers = [
    "Environment :: Console :: Curses",
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "textual>=0.45.0",
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",
    "click>=8.1.0",
    "tomlkit>=0.12.0",
    "python-dateutil>=2.8.0",
    "beautifulsoup4>=4.11.0",
    "pillow>=10.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.20.0",
    "wcwidth>=0.1.7",
    "watchdog>=3.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
    "mypy>=1.0.0",
    "textual-dev>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/tootles-dev/tootles"
"Source" = "https://github.com/tootles-dev/tootles"
"Issues" = "https://github.com/tootles-dev/tootles/issues"

[project.scripts]
tootles = "tootles.cli.main:cli"

[tool.hatch.version]
path = "tootles/__init__.py"

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py38']
line-length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B601"]
